need fixing

+++ app/src/main/java/hu/tb/minichefy/presentation/screens/storage/storage_list/StorageListViewModel.kt
@@ -10,9 +10,8 @@ import hu.tb.minichefy.domain.repository.StorageRepository
 import hu.tb.minichefy.domain.use_case.ValidateNumberKeyboard
 import hu.tb.minichefy.domain.use_case.ValidateQuantity
 import hu.tb.minichefy.domain.use_case.ValidationResult
-import hu.tb.minichefy.presentation.screens.manager.icons.IconManager
 import hu.tb.minichefy.presentation.screens.manager.icons.IconResource
-import hu.tb.minichefy.presentation.screens.manager.icons.ProductIcon
+import hu.tb.minichefy.presentation.screens.manager.icons.FoodIcon
 import hu.tb.minichefy.presentation.ui.theme.SEARCH_BAR_WAIT_AFTER_CHARACTER
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
@@ -33,8 +32,8 @@ class StorageListViewModel @Inject constructor(
     private val _uiState = MutableStateFlow(UiState())
     val uiState: StateFlow<UiState> = _uiState
 
-    private val _editFoodState = MutableStateFlow(ModifyFoodState())
-    val editFoodState: StateFlow<ModifyFoodState> = _editFoodState
+    private val _editFoodState = MutableStateFlow(ModifyFoodQuantityDialogState())
+    val editFoodState: StateFlow<ModifyFoodQuantityDialogState> = _editFoodState
 
     private val _uiEvent = Channel<UiEvent>()
     val uiEvent = _uiEvent.receiveAsFlow()
@@ -62,11 +61,12 @@ class StorageListViewModel @Inject constructor(
         val foodTagList: List<FoodTag> = emptyList(),
         val activeFilter: List<FoodTag> = emptyList(),
         val foodList: List<Food> = emptyList(),
-        val allProductIconList: List<ProductIcon> = IconManager().getAllProductIconList,
+        val allFoodIconList: List<FoodIcon> = FoodIcon.entries,
     )
 
-    data class ModifyFoodState(
-        val foodListPositionIndex: Int = -1,
+    data class ModifyFoodQuantityDialogState(
+        val modifyFoodListPositionIndex: Int = -1,
+        val modifyFood: Food? = null,
         val quantityModifyValue: String = "",
         val unitOfMeasurementModifyValue: UnitOfMeasurement = UnitOfMeasurement.PIECE,
         val isQuantityModifyDialogHasError: Boolean = false,
@@ -79,13 +79,12 @@ class StorageListViewModel @Inject constructor(
 
     sealed class OnEvent {
         data object ClearSearchText : OnEvent()
-        data object ClearEditFoodDialog : OnEvent()
         data object ClearSelectedFoodIndex : OnEvent()
         data object SaveEditFoodQuantities : OnEvent()
         data object SetupEditFoodQuantityDialog : OnEvent()
         data class DeleteFood(val foodId: Long) : OnEvent()
         data class SearchTextChange(val text: String) : OnEvent()
-        data class OnProductClick(val positionIndex: Int) : OnEvent()
+        data class FoodEditButtonClick(val positionIndex: Int) : OnEvent()
         data class FilterChipClicked(val tag: FoodTag) : OnEvent()
         data class ModifyFoodTags(val tag: FoodTag) : OnEvent()
         data class ModifyFoodIcon(val icon: IconResource) : OnEvent()
@@ -95,6 +94,7 @@ class StorageListViewModel @Inject constructor(
 
     fun onEvent(event: OnEvent) {
         when (event) {
+            //page events
             is OnEvent.SearchTextChange -> {
                 viewModelScope.launch {
                     _uiState.update {
@@ -112,10 +112,6 @@ class StorageListViewModel @Inject constructor(
 
             OnEvent.ClearSearchText -> _uiState.update {
                 it.copy(searchText = "")
-            }
-
-            is OnEvent.OnProductClick -> _editFoodState.update {
-                it.copy(foodListPositionIndex = event.positionIndex)
             }
 
             is OnEvent.FilterChipClicked -> {
@@ -142,6 +138,14 @@ class StorageListViewModel @Inject constructor(
                 }
             }
 
+            //edit food events
+            is OnEvent.FoodEditButtonClick -> _editFoodState.update {
+                it.copy(
+                    modifyFoodListPositionIndex = event.positionIndex,
+                    modifyFood = uiState.value.foodList[event.positionIndex]
+                )
+            }
+
             is OnEvent.EditFoodDialogQuantityChange -> {
                 if (validateNumberKeyboard(event.value) == ValidationResult.ERROR) return
 
@@ -158,14 +162,8 @@ class StorageListViewModel @Inject constructor(
                 )
             }
 
-            OnEvent.ClearEditFoodDialog -> _editFoodState.update {
-                it.copy(
-                    quantityModifyValue = "",
-                )
-            }
-
             OnEvent.SetupEditFoodQuantityDialog -> {
-                val editedFood = uiState.value.foodList[editFoodState.value.foodListPositionIndex]
+                val editedFood = editFoodState.value.modifyFood!!
 
                 _editFoodState.update {
                     it.copy(
@@ -194,7 +192,7 @@ class StorageListViewModel @Inject constructor(
                 }
 
                 val updatedFood =
-                    uiState.value.foodList[editFoodState.value.foodListPositionIndex].copy(
+                    editFoodState.value.modifyFood!!.copy(
                         quantity = quantityModifyValue.toFloat(),
                         unitOfMeasurement = editFoodState.value.unitOfMeasurementModifyValue,
                     )
@@ -212,26 +210,30 @@ class StorageListViewModel @Inject constructor(
 
             is OnEvent.ModifyFoodTags -> {
                 viewModelScope.launch {
-                    val modifyFood =
-                        uiState.value.foodList[editFoodState.value.foodListPositionIndex]
-                    if (modifyFood.foodTagList == null) {
+                    val modifyFood = editFoodState.value.modifyFood!!
+                   /* if (modifyFood.foodTagList == null) {
                         storageRepository.saveFoodAndTag(modifyFood.id!!, event.tag.id!!)
-                    } else {
-                        if (modifyFood.foodTagList.contains(event.tag)) {
+                    } else {*/
+                        if (modifyFood.foodTagList != null && modifyFood.foodTagList.contains(event.tag)) {
                             storageRepository.deleteFoodAndTag(modifyFood.id!!, event.tag.id!!)
                         } else {
                             storageRepository.saveFoodAndTag(modifyFood.id!!, event.tag.id!!)
                         }
-                    }
+                    //}
                 }
             }
 
             OnEvent.ClearSelectedFoodIndex -> {
-                _editFoodState.update { it.copy(foodListPositionIndex = -1) }
+                _editFoodState.update {
+                    it.copy(
+                        modifyFoodListPositionIndex = -1,
+                        modifyFood = null
+                    )
+                }
             }
 
             is OnEvent.ModifyFoodIcon -> saveEditedFood(
-                uiState.value.foodList[editFoodState.value.foodListPositionIndex].copy(
+                editFoodState.value.modifyFood!!.copy(
                     icon = event.icon.resource
                 )
             )
